refactor(compare): extract shared types and fmt helpers to _common.py (Phase A 2/N)
Move CheckResult, CrosswalkBuildError, md_escape, and the message-formatting
helpers (fmt_n, fmt_cmp, fmt_cmp_stat — formerly _n, _cmp, _cmp_stat) into
hv_dataqc/compare/_common.py.

The fmt_* helpers currently get called inside check functions to bake
pre-formatted strings into CheckResult.message at check time. That couples
formatting to check logic and undermines Phase E (re-render from JSON
without re-running checks). The proper fix — checks emit structured detail,
renderer formats — is documented as Phase C-prereq in REFACTOR_PLAN.md.
For Phase A the helpers live in _common.py as their interim home; the
import aliases preserve the call sites verbatim so no check code changes.

Also expose THRESHOLDS_PATH publicly from report_io.py (was named
_THRESHOLDS_PATH and referenced in two argparse/main spots that the
previous extraction missed — IDE diagnostics caught them; the regression
test didn't because those code paths don't fire during compare.sh).

Regression-check: COPDGene, ARIC, HCHS all match. Tests: 79/79 pass.
